Fix "error code 22" on some netmirage-core runs
This error was caused by a buffer overrun when parsing rtattr structs
caused by not subtracting the initial struct length from the netlink
message payload size. This would cause RTA_OK to run off the end of
the buffer. Coincidentally due to buffer reuse, that would cause the
code to think that a network interface had two broadcast MAC
addresses (where one was the wrong length). When trying to set the
erroneous double attributes after moving the interface to a new
namespace, the kernel would return an "invalid argument" error.